requirements.txt
rembg
numpy
pillow
tensorflow
keras
fastapi